Publish Version 0.5.0 (#32)
* Removed old cookbook folder.

This folder has been moved to deploy.d

* Resolve typo in container storage deploy.

* Fixed another typo.

* Added odh demo deployment.

* Fixed typo in ocs deployment.;

* Bumped version number.

* Cleanup of cli and container handling.

* Bug fix to odh demo dpeloyment.

* Version bump

* Added version pinning to farosctl command.

* Made many changes to cluster configuration.

In preparation for upcomming changes, much of the config was
streamlined.

* Added bastion hosts interface list injection to cli.

* First pass at router role.

Added a router role that should configure the bastion node to act
as a layer 3 router with firewall. This required quite a bit of
change to the configuration and inventory bits. I expect this has
broken other scripts, but I will find those as I go.

* Added role to create dhcp server on the bastion node/router.

* Added dns server installation to the router node.

* Added ntp-server and router-hardening roles.

This should complete the router creation.
** fingers crossed **

* Updated the farosctl command to manage its own ssh key.

* Cleaned up the data directory initialization to make it more seamless.

* Cleaned up router firewall rules.

* Adjusted router config to use localhost as DNS server.

In order for the router/bastion node to complete the install, it
must use the local DNS server for resoution of the cluster nodes.

* Changed the order for configuring router components.

The LAN bridge must exist before the DHCP and DNS servers can
be created. This means the nodes primary DNS server will not
exist while it is waiting to be created. For this reason, I've
added a secondary DNS server.

* Updated docs for creating router.

* Fixed the shell command in farosctl cli.

* Added configuration option to control ingress network routing.

* The bootstrap script no longer needs to control SSH keys.

This is now handled by the farosctl command.

* Updated create machine command to use the bastion node router.

* Many changes to update the host record deployment to local router.

1) moved create network playbook to apply host-records
2) added necessary configuration params for host-records
3) removed unnecessary config params
4) added wait-for managment-interfaces

* Initial changes to get create cluster to work with internal router.

* Open tftp port.

* open firewall for machine configs.

* Cleaned up create command.

* Removing the concept of providers.

* Updated file structure to have all playbooks in playbook dir.

* Fixed port forwarding to work with direct routing.

* Fixed issue with the loadbalancer ports opening in the wrong zone.

* Don't wait for machine config server to come up.

This service is not garaunteed to be available from the bastion
node and is not specified in the install docs.

* Added option to specify NIC used for install.

* Put management IPs in cockpit rather than hostnames.

This is better when using the router to allow external access.

* Specifying a NIC didn't work....

It seems that all NICs that can DHCP still will. This is undesirable.
For the time being, the easiest way to deal with it is to black hole
any MAC address that I do not want to be active.

* Updated README with BIOS requirements.

* Deleted duplicate become record in handler.

* Fixed bugs in firewall config and dns server config.

* Version Bump

* cleanup of merge artifact.