Changed base image to UBI8.
It appears as if sshpass is no longer required. Because of this, an
unentitled build on UBI 8 is possible. Telnet was also removed to
support this move. Telnet was only used as a debugging tool and never
required for the operation of faros.

Newest Ansible prefers Python 3.8 so that was updated.

Newest PIP prefers to be called as a Python modules rather than
directly. So that was changed. Also updating to the latest pip and wheel
packages to ensure dependencies continue to build nicely.