Added support for static IP addresses on cluster nodes.
A new configuration option has been added to specify the NIC on the node
to use for cluster traffic. If this is not specified, the node will dhcp
boot all interfaces. If this is specified, the node will now configure
that interface with a static ip to eliminate the need to DHCP post
install.

This addresses #9