Changed encryption style to tang instead of tpm.
TPM encryption cipher support can be very hardware dependent and require
very particular BIOS configurations. Additionally, for edge equipment,
it is not a particularly safe option either.

Tang encryption will be used so all the keys can be stored on the
bastion and the installer reliably install encryption on the nodes.

This works toward #150