Static IP Addresses and Bonded Interfaces (#152)
* Added support for static IP addresses on cluster nodes.

A new configuration option has been added to specify the NIC on the node
to use for cluster traffic. If this is not specified, the node will dhcp
boot all interfaces. If this is specified, the node will now configure
that interface with a static ip to eliminate the need to DHCP post
install.

This addresses #9

* Added support for bonding interfaces.

This addresses #129. It is up to the user to ensure the infrastructure
is setup to handle the bond that is configured. That is out of scope of
the installer.